backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
paho-mqtt
python-dotenv
httpx
